[refactor] Enum 타입 String으로 관리 되도록 애노테이션 추가
