+++ src/main/java/com/example/domain/enums/Gender.java
@@ -3,5 +3,4 @@ package com.example.domain.enums;
 public enum Gender {
     MALE, FEMALE, UNKNOWN
 
-
 }
