+++ src/main/java/com/example/domain/enums/StudyState.java
@@ -1,5 +1,5 @@
 package com.example.domain.enums;
 
 public enum StudyState {
-    OPENED, CLOSED
+  RECRUITING, COMPLETED, BEFORE
 }
