+++ src/main/java/com/example/domain/enums/Theme.java
@@ -1,7 +1,5 @@
 package com.example.domain.enums;
 
 public enum Theme {
-
     ON, OFF
-
 }
